Backend/requirements.txt
fastapi
uvicorn
python-socketio
python-engineio
python-dotenv
loguru
pandas
numpy
scikit-learn
python-jose
passlib
python-multipart
aiofiles
httpx
geopy
python-jose
bcrypt
pydantic
email-validator
opencv-python
torch
torchvision
torchaudio 